pyproject.toml
[project]
name = "recommandation-musique"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    # Utilitaires communs
    "jupyter",
    "requests>=2.28.0",
    "python-dotenv>=1.2.1",
    "python-dateutil>=2.8.2",
    "tqdm>=4.66",
    # AWS
    "boto3>=1.42.25",
    "awscli>=1.27.0",
    # Scraping
    "beautifulsoup4",
    # Festival RAG Agent
    "openai>=1.0.0",
    "langchain>=1.2.3",
    "langchain-classic>=0.0.1",
    "langchain-pinecone>=0.2.13",
    "pinecone>=7.3.0",
    "langchain-tools>=0.1.34",
    "langchain-ollama>=1.0.1",
    "langchain-google-genai>=4.1.3",
    # Data / ML (recommandation musicale)
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.11",
    "pyarrow>=14.0",
    "scikit-learn>=1.3",
    "implicit>=0.7",
    "zstandard>=0.22",
    # API
    "fastapi>=0.135.1",
    "uvicorn[standard]>=0.42.0",
    "websockets>=12.0",
    "pydantic>=2.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
airflow = [
    "apache-airflow>=2.5.0",
    "apache-airflow-providers-amazon>=8.0.0",
]
spark = [
    "pyspark>=3.3.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
